requirements.txt
sentence-transformers
langchain
streamlit
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
google-generativeai
langchain-google-genai
-e .